Fix scopes, remove most logs
Made all variables and functions that were only beeing used in their own
file not accessible elsewhere. Removed most of the Serial.print's as
they seemed to be causing errors after setting the internal clock. For
some reason, the IDE is requesting me to include the used libraries in
the main header, not only in the header for the file that uses them.